backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
numpy==2.2.3
Pillow==11.1.0
requests==2.32.3
gunicorn==23.0.0
python-dotenv==1.0.1
onnxruntime==1.20.1
huggingface_hub>=0.20.0
